requirements.txt
Django>=1.8
psycopg2>=2.5.2
djangorestframework>=3.2.3
